Metoda do tworzenia zleceń Zmiana typu wagi z inta na decimal

+++ Warehouse/Models/Custom/OrderModels/OrdersPositions.cs
@@ -9,9 +9,9 @@ namespace Warehouse.Models.Custom
         public string Name { get; set; }
         public int Order_id { get; set; }
         public int Amount { get; set; }
-        public int Weight_Gross { get; set; }
+        public decimal Weight_Gross { get; set; }
         public int? Amount_Received { get; set; }
-        public int? Weight_Gross_Received { get; set; }
+        public decimal? Weight_Gross_Received { get; set; }
         public string Created_At { get; set; }
         public string Edited_At { get; set; }
     }
